Issue 139,137,124,188 all change access popup functionality script implemented all for catalogs
