+++ src/main/java/aicore/pages/ChangeAccessPopUpPage.java
@@ -1,0 +1,11 @@
+package aicore.pages;
+
+import com.microsoft.playwright.Page;
+
+public class ChangeAccessPopUpPage extends AbstractChangeAccessPopUpPage {
+
+	public ChangeAccessPopUpPage(Page page) {
+		this.page = page;
+	}
+
+}
