+++ src/main/java/aicore/utils/ModelPageUtils.java
@@ -425,11 +425,13 @@ public class ModelPageUtils {
 			addMemberSearchbar.click();
 			addMemberSearchbar.fill(username);
 			page.waitForTimeout(2000);
-			addMemberSearchbar.press("ArrowDown");
-			addMemberSearchbar.press("Enter");
+//			addMemberSearchbar.press("ArrowDown");
+//			addMemberSearchbar.press("Enter");
+			page.getByTitle(username).click();
+			page.click(RADIO_BUTTON_XPATH.replace("{role}", role));
+			page.click(SAVE_BUTTON_XPATH);
 		}
-		page.click(RADIO_BUTTON_XPATH.replace("{role}", role));
-		page.click(SAVE_BUTTON_XPATH);
+
 		// THESE ELEMENTS REMOVED ON A SEMOSS UPDATE ON May 12, 2025
 //		page.click(MEMBER_ADDED_SUCCESS_TOAST_MESSAGE_CLOSE_ICON_XPATH);
 //		page.locator(MEMBER_ADDED_SUCCESS_TOAST_MESSAGE_XPATH)
